chore: re-organize files, fix KG issues, and test TS functions
