--- client/src/features/project/ProjectEnums.ts
@@ -1,11 +1,0 @@
-export enum ProjectIndexingStatuses {
-  InProgress = "in-progress",
-  Success = "success",
-  Failure = "failure",
-}
-
-export enum MigrationStartScopes {
-  All,
-  OnlyTemplate,
-  OnlyVersion,
-}
