+++ client/src/features/project/index.ts
@@ -2,8 +2,8 @@ import ProjectDatasetsListView from "./dataset/ProjectDatasetsListView";
 import ProjectDatasetsView from "./dataset/ProjectDatasetsView";
 import ProjectDatasetShow from "./dataset/ProjectDatasetShow";
 import ProjectEntityHeader from "./components/ProjectEntityHeader";
-import ProjectKnowledgeGraphStatus from "./ProjectKnowledgeGraphStatus";
-import { ProjectFileLineage, ProjectFileView } from "./file";
+import ProjectKnowledgeGraphStatus from "./components/ProjectKnowledgeGraphStatus";
+import { ProjectFileLineage, ProjectFileView } from "./components/File";
 
 export { ProjectEntityHeader };
 export { ProjectDatasetsListView, ProjectDatasetShow, ProjectDatasetsView };
