+++ client/src/features/project/projectEnums.ts
@@ -1,0 +1,48 @@
+/*!
+ * Copyright 2023 - Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export enum ProjectIndexingStatuses {
+  InProgress = "in-progress",
+  Success = "success",
+  Failure = "failure",
+}
+
+export enum MigrationStartScopes {
+  All,
+  OnlyTemplate,
+  OnlyVersion,
+}
+
+export enum ProjectMigrationLevel {
+  Level1 = "1-all-good", // ? success
+  Level2 = "2-some-info-no-update", // ? info
+  Level3 = "3-minor-update", // ? info
+  Level4 = "4-suggested-update", // ? warning
+  Level5 = "5-required-update", // ? danger
+  LevelE = "error", // ? danger
+  LevelX = "x-unknown", // ? danger
+}
+
+// ! Double check if it's actually used
+export enum ProjectIndexingLevel {
+  Level1 = "1-all-good", // ? success
+  Level2 = "2-some-info-no-update", // ? info
+  Level5 = "5-required-update", // ? danger
+  LevelE = "error", // ? danger
+  LevelX = "x-unknown", // ? danger
+}
